// src/shared/api/types/gen/trade.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * ViaTrade API
 * API для платформы инвестиционного анализа ViaTrade
 * OpenAPI spec version: v1
 */
import type { TradeSignal } from './tradeSignal';

export type Trade = {
	id: number;
	dateOpen: string;
	dateClose?: string;
	tradeOpen: number;
	tradeClose?: number;
	netIncome?: number;
	count: number;
	price: number;
	tradeTypeId: number;
	tradeCodeId: number;
	userId: number;
	tradeSignal: TradeSignal;
};